ui: create unique component only for public key editing
